requirements.txt
openai>=2.26.0
python-dotenv>=1.2.2
PyAutoGUI>=0.9.54
PyGetWindow>=0.0.9
Pillow>=12.1.1
playwright>=1.58.0
pytesseract>=0.3.13
pyperclip>=1.11.0
pywinauto>=0.6.9
